pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.12.1",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "telegram>=0.0.1",
    "yookassa>=3.5.0",
]
